v0.0.4
First version of rudimentary game viewer, which allows for
navigating through the moves of a specific game.

This initial version is doing the processing and rendering in
the backend and serving static pages, but for v2 I'd like to
build a JS front-end to display this stuff in a much smoother
manner, and use far fewer page calls.